cambio las coordendas de los barcos a int

+++ backend/src/logica/entidades/Vehiculo.java
@@ -5,8 +5,8 @@ import com.google.gson.JsonArray;
 public abstract class Vehiculo {
 
 	private int id;
-	private float x;
-	private float y;
+	private int x;
+	private int y;
 	private float velocidad;
 	private float velocidadAngular;
 	private String spriteVivo;
@@ -17,7 +17,7 @@ public abstract class Vehiculo {
 		
 	}
 	
-	public Vehiculo(int id, float x, float y, float velocidad, float velocidadAngular, String spriteVivo, String spritesLaterales, float initialRotation) {
+	public Vehiculo(int id, int x, int y, float velocidad, float velocidadAngular, String spriteVivo, String spritesLaterales, float initialRotation) {
 		super();
 		this.id = id;
 		this.x = x;
@@ -41,7 +41,7 @@ public abstract class Vehiculo {
 		return x;
 	}
 	
-	public void setX(float x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 	
@@ -49,7 +49,7 @@ public abstract class Vehiculo {
 		return y;
 	}
 	
-	public void setY(float y) {
+	public void setY(int y) {
 		this.y = y;
 	}
